Update README.md
his assignment is a quick assignment to get you started with programming in C, as well as compiling, linking, running, and debugging. Your task is to write 5 small C programs. Your program must follow the input-output guidelines listed in each section exactly, with no additional or missing output.

No cheating or copying will be tolerated in this class. Your assignments will be automatically checked with plagiarism detection tools that are pretty powerful. Hence, you should not look at your friend’s code. See CS department’s academic integrity policy at: http://academicintegrity.rutgers.edu/

First: Prime Numbers (5 Points)

You have to write a program that will read an array from a le and print if the numbers in the le are right truncatable primes. A right truncatable prime is a prime number, where if you truncate any numbers from the right, the resulting number is still prime. For example, 3797 is a truncatable prime number number because 3797, 379, 37, and 3 are all primes.

Input-Output format: Your program will take the le name as input. The rst line in the le provides the total number of values in the array. The subsequent lines will contain an integer value. For example a sample input le \ le1.txt” is: